reduce fetch params creation logic
