+++ src/json_event.ts
@@ -1,31 +1,25 @@
-import type { DispatchParams } from "./type_flyweight.js";
+import type { DispatchParams, FetchParamsInterface } from "./type_flyweight.js";
 
-import { getRequestParams, createRequest } from "./type_flyweight.js";
+import { getRequestParams, createFetchParams } from "./type_flyweight.js";
 import { setThrottler, getThrottleParams, shouldThrottle } from "./throttle.js";
 import { getQueueParams, enqueue } from "./queue.js";
 
-interface JsonRequestInterface {
-	request: Request;
-	action: string;
-	abortController: AbortController;
-}
-
-interface JsonRequestQueuedInterface extends JsonRequestInterface {
+interface JsonRequestQueuedInterface extends FetchParamsInterface {
 	status: "queued";
 	queueTarget: EventTarget;
 }
 
-interface JsonRequestRequestedInterface extends JsonRequestInterface {
+interface JsonRequestRequestedInterface extends FetchParamsInterface {
 	status: "requested";
 }
 
-interface JsonRequestResolvedInterface extends JsonRequestInterface {
+interface JsonRequestResolvedInterface extends FetchParamsInterface {
 	status: "resolved";
 	response: Response;
 	json: any;
 }
 
-interface JsonRequestRejectedInterface extends JsonRequestInterface {
+interface JsonRequestRejectedInterface extends FetchParamsInterface {
 	status: "rejected";
 	error: any;
 }
@@ -50,25 +44,16 @@ export class JsonEvent extends Event implements JsonEventInterface {
 }
 
 export function dispatchJsonEvent(dispatchParams: DispatchParams) {
-	let requestParams = getRequestParams(dispatchParams);
-	if (!requestParams) return;
-
 	// this could be one function
 	// let throttleParams = getThrottleParams(dispatchParams);
 	// if (shouldThrottle(dispatchParams)) return;
 
-	let abortController = new AbortController();
-
 	// setThrottler(dispatchParams, throttleParams, abortController);
 
-	let { action } = requestParams;
-	let request = createRequest(dispatchParams, requestParams, abortController);
-
-	let fetchParams: JsonRequestInterface = {
-		action,
-		request,
-		abortController,
-	};
+	// this can be one function
+	// createFetchParams(dispatchParams, requestParams)
+	let fetchParams = createFetchParams(dispatchParams);
+	if (!fetchParams) return;
 
 	// let queueParams = getQueueParams(dispatchParams);
 	// if (queueParams) {
@@ -91,7 +76,7 @@ export function dispatchJsonEvent(dispatchParams: DispatchParams) {
 }
 
 function fetchJson(
-	fetchParams: JsonRequestInterface,
+	fetchParams: FetchParamsInterface,
 	dispatchParams: DispatchParams,
 ): Promise<void> | undefined {
 	if (fetchParams.request.signal.aborted) return; // maybe?
